requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.2.0
autopep8==2.3.1
av==11.0.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.43.1
black==24.3.0
blinker==1.7.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
ctranslate2==4.1.0
chromadb==0.4.24
cycler==0.12.1
dataclasses-json==0.6.4
deepdiff==6.7.1
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
ecdsa==0.18.0
emoji==2.10.1
et-xmlfile==1.1.0
fake-useragent==1.5.0
fastapi==0.110.0
faster-whisper==1.0.1
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.7
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
gradio==4.37.2
gradio_client==1.0.2
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.0
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonify==0.5
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.38
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.85
litellm==1.30.7
lxml==5.1.0
mammoth==1.8.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.13.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.2
mdurl==0.1.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.17.1
openai==1.35.14
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
oracledb==1.4.2
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
passlib==1.7.4
pathspec==0.12.1
peewee==3.17.1
pillow==10.2.0
platformdirs==4.2.0
posthog==3.5.0
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycodestyle==2.12.0
pycparser==2.21
pydantic
pydantic_core
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.7
pymupdf4llm==0.0.9
PyMuPDFb==1.24.6
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-iso639==2024.2.7
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.1
pytz==2024.1
pyxlsb==1.0.10
PyYAML==6.0.1
rapidfuzz==3.6.2
rapidocr-onnxruntime==1.2.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruff==0.5.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentence-transformers==2.5.1
setuptools==69.2.0
shapely==2.0.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.36.0
sympy==1.12
tabulate==0.9.0
tavily-python==0.3.4
tbb==2021.13.0
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.7.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
transformers==4.38.2
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
uuid==1.30
uvicorn==0.28.0
watchdog==4.0.1
watchfiles==0.21.0
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
yarl==1.9.4
zipp==3.18.1
langid==1.1.6
langchain-ollama== 0.1.1
langchain-community==0.2.5
llama_index==0.10.64
llama_index.utils.workflow==0.1.0
streamlit_modal==0.1.2
